+++ prettier.config.js
@@ -3,7 +3,6 @@ export default {
   semi: false,
   singleQuote: true,
   trailingComma: 'all',
-  pluginSearchDirs: false,
   plugins: ['@ianvs/prettier-plugin-sort-imports'],
   importOrder: ['^@', '^[a-zA-Z0-9-]+', '^[./]'],
 }