+++ src/index.ts
@@ -28,6 +28,8 @@ import { spliceChangesIntoString, visit } from './utils.js'
 
 let base = await loadPlugins()
 
+const ESCAPE_SEQUENCE_PATTERN = /\\(['"\\nrtbfv0-7xuU])/g
+
 function createParser(
   parserFormat: string,
   transform: (ast: any, context: TransformerContext) => void,
@@ -140,6 +142,44 @@ function transformDynamicAngularAttribute(attr: any, env: TransformerEnv) {
           },
         }),
       })
+    },
+
+    TemplateLiteral(node, path) {
+      if (!node.quasis.length) return
+
+      let concat = path.find((entry) => {
+        return (
+          entry.parent &&
+          entry.parent.type === 'BinaryExpression' &&
+          entry.parent.operator === '+'
+        )
+      })
+
+      for (let i = 0; i < node.quasis.length; i++) {
+        let quasi = node.quasis[i]
+
+        changes.push({
+          start: quasi.start,
+          end: quasi.end,
+          before: quasi.value.raw,
+          after: sortClasses(quasi.value.raw, {
+            env,
+
+            // Is not the first "item" and does not start with a space
+            ignoreFirst: i > 0 && !/^\s/.test(quasi.value.raw),
+
+            // Is between two expressions
+            // And does not end with a space
+            ignoreLast:
+              i < node.expressions.length && !/\s$/.test(quasi.value.raw),
+
+            collapseWhitespace: {
+              start: concat?.key !== 'right' && i === 0,
+              end: concat?.key !== 'left' && i >= node.expressions.length,
+            },
+          }),
+        })
+      }
     },
   })
 
@@ -454,6 +494,7 @@ function sortStringLiteral(
   node: any,
   {
     env,
+    removeDuplicates,
     collapseWhitespace = { start: true, end: true },
   }: {
     env: TransformerEnv
@@ -463,43 +504,44 @@ function sortStringLiteral(
 ) {
   let result = sortClasses(node.value, {
     env,
+    removeDuplicates,
     collapseWhitespace,
   })
+
   let didChange = result !== node.value
+
+  if (!didChange) return false
+
   node.value = result
 
-  // A string literal was escaped if:
-  // - There are backslashes in the raw value; AND
-  // - The raw value is not the same as the value (excluding the surrounding quotes)
-  let wasEscaped = false
+  // Preserve the original escaping level for the new content
+  let raw = node.extra?.raw ?? node.raw
+  let quote = raw[0]
+  let originalRawContent = raw.slice(1, -1)
+  let originalValue = node.extra?.rawValue ?? node.value
 
   if (node.extra) {
-    // JavaScript (StringLiteral)
-    wasEscaped =
-      node.extra?.rawValue.includes('\\') &&
-      node.extra?.raw.slice(1, -1) !== node.value
-  } else {
-    // TypeScript (Literal)
-    wasEscaped =
-      node.value.includes('\\') && node.raw.slice(1, -1) !== node.value
-  }
+    // The original list has ecapes so we ensure that the sorted list also
+    // maintains those by replacing backslashes from escape sequences.
+    //
+    // It seems that TypeScript-based ASTs don't need this special handling
+    // which is why this is guarded inside the `node.extra` check
+    if (originalRawContent !== originalValue && originalValue.includes('\\')) {
+      result = result.replace(ESCAPE_SEQUENCE_PATTERN, '\\\\$1')
+    }
 
-  let escaped = wasEscaped ? result.replace(/\\/g, '\\\\') : result
-
-  if (node.extra) {
     // JavaScript (StringLiteral)
-    let raw = node.extra.raw
     node.extra = {
       ...node.extra,
       rawValue: result,
-      raw: raw[0] + escaped + raw.slice(-1),
+      raw: quote + result + quote,
     }
   } else {
     // TypeScript (Literal)
-    let raw = node.raw
-    node.raw = raw[0] + escaped + raw.slice(-1)
+    node.raw = quote + result + quote
   }
-  return didChange
+
+  return true
 }
 
 function isStringLiteral(node: any) {
@@ -513,6 +555,7 @@ function sortTemplateLiteral(
   node: any,
   {
     env,
+    removeDuplicates,
     collapseWhitespace = { start: true, end: true },
   }: {
     env: TransformerEnv
@@ -530,6 +573,7 @@ function sortTemplateLiteral(
 
     quasi.value.raw = sortClasses(quasi.value.raw, {
       env,
+      removeDuplicates,
       // Is not the first "item" and does not start with a space
       ignoreFirst: i > 0 && !/^\s/.test(quasi.value.raw),
 
@@ -553,6 +597,7 @@ function sortTemplateLiteral(
           ignoreFirst: i > 0 && !/^\s/.test(quasi.value.cooked),
           ignoreLast:
             i < node.expressions.length && !/\s$/.test(quasi.value.cooked),
+          removeDuplicates,
           collapseWhitespace: collapseWhitespace && {
             start: collapseWhitespace && collapseWhitespace.start && i === 0,
             end:
@@ -629,6 +674,10 @@ function isSortableCallExpression(
   return false
 }
 
+// TODO: The `ast` types here aren't strictly correct.
+//
+// We cross several parsers that share roughly the same shape so things are
+// good enough. The actual AST we should be using is probably estree + ts.
 function transformJavaScript(
   ast: import('@babel/types').Node,
   { env }: TransformerContext,
@@ -946,7 +995,7 @@ function transformSvelte(ast: any, { env, changes }: TransformerContext) {
           env,
           ignoreFirst: i > 0 && !/^\s/.test(value.raw),
           ignoreLast: i < attr.value.length - 1 && !/\s$/.test(value.raw),
-          removeDuplicates: false,
+          removeDuplicates: true,
           collapseWhitespace: false,
         })
         value.data = same
@@ -955,7 +1004,7 @@ function transformSvelte(ast: any, { env, changes }: TransformerContext) {
               env,
               ignoreFirst: i > 0 && !/^\s/.test(value.data),
               ignoreLast: i < attr.value.length - 1 && !/\s$/.test(value.data),
-              removeDuplicates: false,
+              removeDuplicates: true,
               collapseWhitespace: false,
             })
       } else if (value.type === 'MustacheTag') {
@@ -1113,11 +1162,22 @@ export const parsers: Record<string, Parser> = {
     staticAttrs: ['class', 'className'],
   }),
 
+  hermes: createParser('hermes', transformJavaScript, {
+    staticAttrs: ['class', 'className'],
+  }),
+
   typescript: createParser('typescript', transformJavaScript, {
     staticAttrs: ['class', 'className'],
   }),
 
   'babel-ts': createParser('babel-ts', transformJavaScript, {
+    staticAttrs: ['class', 'className'],
+  }),
+
+  oxc: createParser('oxc', transformJavaScript, {
+    staticAttrs: ['class', 'className'],
+  }),
+  'oxc-ts': createParser('oxc-ts', transformJavaScript, {
     staticAttrs: ['class', 'className'],
   }),